:hammer: a little optimization of style and others
